requirements.txt
# Chess Fairplay Analyzer Dependencies
click>=8.1.0
requests>=2.28.0
python-chess>=1.999
pyyaml>=6.0
jinja2>=3.1.0
tqdm>=4.65.0
colorama>=0.4.6
numpy>=1.24.0

# Data Analysis and Visualization
pandas>=1.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
reportlab>=4.0.0
openpyxl>=3.10.0